Refactor filter type definitions and usage
Moved `FilterKey` and `CheckedFilters` types to relevant hooks for better modularity. Updated filter handling logic to align with the new type locations, improving code clarity and maintainability.
