--- packages/packages/core/editor-global-classes/src/components/search-and-filter/types.ts
@@ -1,7 +1,0 @@
-export type FilterKey = 'empty' | 'onThisPage' | 'unused';
-
-export type CheckedFilters = {
-	empty: boolean;
-	onThisPage: boolean;
-	unused: boolean;
-};
