+++ packages/packages/core/editor-global-classes/src/hooks/use-filtered-css-class-usage.tsx
@@ -2,9 +2,10 @@ import { useMemo } from 'react';
 import { __useActiveDocument as useActiveDocument } from '@elementor/editor-documents';
 
 import { type CssClassUsageContent, type EnhancedCssClassUsage } from '../components/css-class-usage/types';
-import { type FilterKey } from '../components/search-and-filter/types';
 import { useCssClassUsage } from './use-css-class-usage';
 import { useAllCssClassesIDs, useEmptyCssClass } from './use-empty-css-class';
+
+export type FilterKey = 'empty' | 'onThisPage' | 'unused';
 
 type FilteredCssClassUsage = Record< FilterKey, string[] >;
